sparc64 work. 1. Formatting fixes. 2. When writing to DMMU ASI's, simple membar() can be used in place of flush(). 3. Substantial changes in the way the TLB is taken over. 4. Remove unneeded functions.
This is the first revision that also runs on a real world Ultra 5 with UltraSPARC IIi
processor.

Note that 3. needs further work as the current implementation depends on the fact
that the compiler will use registers for local variables in take_over_tlb_and_tt().
Rewrite of that function into assembly is to follow.