Extend the OFW device tree node with a void pointer. The boot loader is supposed to reset all pointers in all nodes. The kernel uses the pointer to recognize that a particular device has already been visited and initialized.
Read interrupt mapping from the OFW device tree for FHC devices (z8530) and
EBUS devices (ns16550). In case of FHC devices, remove hardwired values from
the code and use only values read from the tree. FHC initialization is started
when a FHC device wants to map its interrupt. In case of EBUS devices, map
the interrupt to INO. Interrupt enabling in the interrupt controller for that
specific interrupt is not implemented yet.