Cleanup the waitq_wakeup() interface. Replace numeric constants (i.e. 0) and boolean constants (i.e. false) with more readable WAKEUP_FIRST. Also change the type of the second argument of waitq_wakeup() to a newly introduced type wakeup_mode_t.
Fix behaviour of waitq_wakeup() in case
that WAKEUP_ALL semantics is required
but no threads are sleeping in the wait
queue. This is a similar fix to that of
Jan Hudecek committed in the RCU branch,
but, IMHO, is more straightforward and
also doesn't eat up previous missed
wakeups.