Introduce a more platform-neutral name for programmed I/O.
The new API looks like pio_read_n() or pio_write_n(), where n is 8, 16 or 32.
The old API (i.e.  inb(), inw(), inl(), outb() outw(), outl()) may have made
some people think that the interface is only to be used with the separate I/O
space. That's not the case. This API is to be implemented on all platforms
so that we can finally have really generic kernel device drivers.