Do not start devman automatically.
There seems to be a problem when both devman and netstart are run. Either dp8390
crashes while writing to an IO port (when netstart runs after devman) or devman
hangs (when netstart runs before devman).  The problem appears to be related to
the fact that both the dp8390 and com2 try to pio_enable() ports from the same
I/O range.