Assert replaced with condition
The assert was originally there for debugging purposes. However, there is
no guarantee that the console will create the callback phone immediatelly.
It may be busy doing other things and impatient user may be already tapping
the keyboard. It is definitely better to discard some keys than to crash
the driver ;-).