Libposix functions are without posix_ prefix
Prior this commit, libposix headers declared all functions as posix_*
and used macros to rename e.g. strncpy to posix_strncpy in all (ported)
sources.

After this change, libposix headers look as normal POSIX compliant headers
(well, almost) and no renaming is done in the source codei (of the ported
applications). Instead, the renaming is done at object files level to
bypass weird problems that are bound to happen if you use macros.

The scheme is following. libposix headers use special macro to declare
the names. When included from outside, the functions have their normal
(standard) names. When included from the libposix sources, posix_ prefix
is added. Thus, when libposix is compiled and linked, it contains the
posix_* naming while compiling of ported software uses the normal
non-prefixed versions. This way the posix_* can use HelenOS libc without
any problem. Before linking, the posix_* prefix is removed from all
symbols and special prefix __helenos_libc_ is added to all functions
that exists in our (HelenOS) libc and its name clashes with the POSIX
one.

The following happens, for example, to the open() function that exists in
both libposix and in libc.
 - Headers and sources of libc are left intact.
 - Copy of libc.a is made and to all clashing functions is added the
   __helenos_libc prefix. This library is called libc4posix.a.
 - __POSIX_DEF__(open)(const char *) is used in libposix headers. This
   macro expands to plain open when included from the "outside world".
   But it expands to posix_open when included from libposix sources.
 - Libposix is compiled and linked, containing posix_open() that
   internally calls open() [the original one from libc].
 - Libposix is transformed - all open() are replaced with prefix variant:
   __helenos_libc_open() and all posix_open() are replaced with open().
   The transformed library is stored as libposixaslibc.a

Binutils and PCC are then linked with libc4posix and libposixaslibc
libraries instead of libc and libposix as was done previously.

WARNING: it looks that binutils, PCC and MSIM still works but not all
architectures were tested.