Explicitly treat the str argument to fat_lfn_str_nlength() as unaligned.
fat_lfn_str_nlength() was defined to take uint16_t * and as such had
reasons to assume natural alignment of the pointed object. This was
however not the case when the function was called from fat_lfn_size()
because the argument there is an address from inside a packed structure.