Fix race between ipc_answerbox_slam_phones() and ipc_cleanup()
We modify _ipc_call() to support sending of pre-forgotten calls. This is
handy in ipc_answerbox_slam_phones() when after slamming a client's
phone, we may no longer be sure of the client's continued existence.

A pre-forgotten call is delivered to the callee in an already forgotten
state, so that the caller is no longer expected to wait for it nor it is
expected to be still around. The caller is held alive for the necessary
time, but can disappear as soon as we release it.