Partially revive support for mount table listing and walking
This implementation is purely client-based. Using a simple file system
traversal, the library figures out what are the actual mountpoints. It
only discovers mountpoints that it can reach from its VFS root.

The benefit is that this works and produces reasonable results even in
the environment of per-client VFS roots, whereas a global mount table
would present paths that might not make any sense to the client
(notwithstanding the fact that VFS no longer knows the mountpoint
paths).

Listing of file system names is still broken as of this commit.