Revert the order of link and unlink operations during rename
This changeset reverts the order of link and unlink operations that
occur within a single rename operation to be link first and unlink
second.

The file system implementations are not ready for the opposite ordering.

For example ext4fs decrements the the link count by 2 in total when
removing the last non-dot name of a directory, but increments only by 1
when adding the new name. This results in an assertion failure on the
next unlink, because the linkcount gets out of ballance.

In addition, this changeset fixes a node reference leak in error paths.