usbhost: introduce dma_buffer
Dma buffers shall simplify management of bounce buffers, eliminating the
errors of using malloc/malloc32, avoiding addr_to_phys.

The alloc function can be passed a policy to influence alignment and
flags. This will be used in future by xhci to allow bounce buffers
outside the lower 32b of addressable space, if the hardware supports it.

Also, the buffers are written with particular extension in mind: memory
pools. In previous HCs, there are a lot of situations where
significantly smaller buffer then PAGE_SIZE is needed, yet a page needs
to be allocated. We could solve this issue by building an allocator on
top of dma-available memory.