Adapt to the improved sun4u model in QEMU 2.11
In QEMU 2.11 the sun4u machine will become much more like the real world
Ultra 5. This means that the EBUS with the serial will move behind a
PCI-to-PCI bridge. Our PCI driver can handle this situation well, but we
need to change the boot argument which specifies the path to the serial
accordingly.