usbhost bus: refactor the bus ops
This way, method names better represent the entity it is working with.
Their semantics was shifted a bit.

Regarding the tree of structures:
    bus <- device <- endpoint <- batch
Previously, devices were kept in DDF function nodes, and endpoints had
pointer to the bus and device. Now, devices have pointer to bus,
endpoints don't.

Pointer to hcd_t in bus is WIP, and will be removed.