tmon: refactoring
Tests now live as descendants of a common interface, which allows them to read generic parameters and be executed on diagnostic devices. All tests have a common main function, which calls this interface, and poses as a command to the main function of the tmon utility. Stress tests have therefore been moved to a separate file.