usbdiag: refactoring
Reconciliated the `usb_diag` and `usbdiag` prefix in favor of the latter. Removed a ton of copy-pasta code for the burst tests in favor of single generic static burst test for every direction, which is then called by individual functions for different endpoint type. This also allowed to get rid of the ugly translation macros in `device.c`. Deleted unused `usbdiag.h` header file. Refactored the remote interface to comply with all aforementioned modifications.