Use `errno_t` in all uspace and kernel code.
Change type of every variable, parameter and return value that holds an
<errno.h> constant to either `errno_t` (the usual case), or `sys_errno_t`
(some places in kernel). This is for the purpose of self-documentation,
as well as for type-checking with a bit of type definition hackery.

After this commit, HelenOS is free of code that mixes error codes with non-error
values on the assumption that error codes are negative.