dltest needs test that calls a function
During bringup of dynamic linking on a platform it is difficult to test
dynamic function calls (via PLT) if the binaries (including dltest)
are statically linked. We can do that by adding a function to libdltest
that calls another function. Since we don't use -Bsymbolic, the call
will be made via PLT.