Add nonstandard ftell64() and fseek64().
Currently, the C standard does not define any functions that can access file
offset with full 64-bit range. Standard ftell()/fseek() are nonportable on files
over 2 GB, while fgetpos()/fsetpos() can work with arbitrary length file, but
keep the position opaque, and no cannot be used for seeks to known numerical
offsets.

POSIX defines ftello()/fseeko(), which _can_ be 64-bit, but aren't guaranteed
to be by the standard, and come with additional baggage. Also, the naming
is rather awkward.

In short, there is no portable interface for this functionality, so
a pair of nonstandard functions is an adequate answer for this.