Travis CI improvements (#16)
Several improvements of the CI process. The dashboard is less cluttered as less variables are set. Ported software is built too. But only for ia32 as that seems to be the most healthy platform in this context. Overall, this should provide shorter feedback than the nightly builds.

Due to complexity of the building process (especially when ported software is compiled) a separate script is used with Travis CI to control the build. See comments in .travis.yml for details how to customize what is being built (if you plan to use it on your feature branches).