usb: update copyrights
The data was generated by a script, guided manually. If you feel your
name is missing somewhere, please add it!

The semi-automated process was roughly:
  1) Changes per file and author (limited to our team) were counted
  2) Trivial numbers were thrown away
  3) Authors were sorted by lines added to file
  4) All previous copyrights were replaced by the newly generated one
  5) Hunks changing only year were discarded

It seems that a lot of my copyrights were added. It is due to me being
both sticking my nose everywhere and lazy to update the copyright right
away :)