Switch completely to HelenOS-specific toolchain target (#44)
As previously discussed, utilizing the compiler-provided headers and libraries would
allow us to avoid a number of problems and reduce the amount of architecture-support
code that we have to implement ourselves. With this change, the support for using
Linux-targeted toolchain, or even the host toolchain itself, is removed, due to the
inherent issues caused by such setup.

Additionally, the toolchain installation is a bit simplified to install all targets into a single
prefix. This prefix can be added to the PATH so that one can simply execute
e.g. amd64-helenos-gcc instead of needing to use the whole path.