Download travis toolchain from Amazon S3 for now
Also a few minor changes to the script.
The toolchain I uploaded is not static. Instead it's built in virtualized
Ubuntu Trusty using a modified toolchain.sh with size optimizations enabled.
The resulting archive is a lot smaller than the static-link version.
The modified script is downloadable from the same path as the archives
(named "toolchain.sh" as usual).

Ask me for write access to the bucket if you need/want it.