Make INTERFACE_LOC_SUPPLIER serial again
This commit entirely reverts commit cbdb38ff8f. The interface
unfortunately cannot be parallel/cloned due to the fact that each loc
supplier connection is stateful and is expected to begin with a
LOC_SERVER_REGISTER call.  Failing that, cloned exchanges to the
location service supplier interface immediately fail with EREFUSED,
because the method mismatches.

This results in various sporadic failures to register services or add
services into categories, which in turn results in symptoms such as
non-functioning keyboard and mouse input.

If memory serves well, the issue that cbdb38ff8f addressed was that
category change events can transitively result in new requests to the
interface over the same phone. Around the time of cbdb38ff8f there used
to be a hard limit on the number of concurrent async calls and calls
that exceeded this limit got blocked. This is no longer the case, so the
loc supplier fibril is not waiting for any of its own IPC. Hence
reverting cbdb38ff8f should be safe.