correcting find_command() which returns allocated strings
The old version of find_command() relied on a global variable called
'found'. This variable had to be released after every call on
find_command(). The old version did this immediatly after the call.

The code directly implied that this has to be changed with the following
comment:
'FIXME: Just have find_command() return an allocated string'

This commit removes the global variable, changes the implementation
of find_command() to return an allocated string and adds/removes
certain str_dup() calls to match the new implementation