Add can_cast_size_t_to_int
There are already several places where we check that it is okay to cast
size_t value to int. This function should unify these places to replace
non-trivial code with a simple function call.

Also added a (hopefully correct) test for this function.

More functions should be added on-demand basis.