Correcting test of the comparison function
This commit removes the old file description with qsort
and gnome sort. Furthermore it corrects how the return
value from the comparison function gets handled. In the
previous version the comparison function must have returned
-1, 0 or 1. The updated version will accept values smaller
than -1 too. This change was needed since the documentation
does not specify it and because all other *cmp functions
tend to return values smaller than -1 (as example strcmp())