Fix version 4 UUID generation
According to the standard, the 4 most significant bits of b[6]
are to be 0x4. Our code accidentally cleared the wrong portion
of the byte.
Also the most significant bit of b[8] should be 1 and the second
most significant should be 0. Our code had them swapped.

Found by Matthieu Riolo, thanks!