This commit fixes #684 and adds autocompletion for subdirectories in search_path. The original implementation of find_command() will find cmd inside the working directory even without the prefix './'. However, those cmds would have not been visible in the auto completion. This commit prevents this from happening.
The problem can be traced back to vfs_absolutize() which treats paths
without the './' prefix as relative paths. The changes made unsures
that only relative or absolute pathes will be passed to vfs_absolutize()

This commit also fixes the problem, that subdirectories inside the
search_path would not be autocomplete. BDSH would execute such paths.
The reason being is that such paths was interpreted as relative paths.
Now BDSH makes a distinction between relatives paths or paths inside
search_path