Add support for serial kernel console on sun4u
This feature is currently disabled by default (CONFIG_SUN_TTY) due to
the following behavior observed with QEMU 3.1. After producing a couple
of lines of serial output, the kernel seems to hang on an IO read of the
NS 16550's LSR register (which it needs in order to test for
LSR_TH_READY). Providing some input on the serial console results in
unblocking the kernel for a while until it blocks again a couple of
lines of output later.