Change type of uspace pointers in kernel from pointer type to numeric type
From kernel's perspective, userspace addresses are not valid pointers,
and can only be used in calls to `copy_to/from_uspace()`.
Therefore, we change the type of those arguments and variables to
`uspace_addr_t` which is an alias for `sysarg_t`.

This allows the compiler to catch accidental direct accesses to
userspace addresses.

Additionally, to avoid losing the type information in code,
a macro `uspace_ptr(type)` is used that translates to `uspace_addr_t`.
I makes no functional difference, but allows keeping the type information
in code in case we implement some sort of static checking for it in the future.