Improvement alias/unlias according to feedback
This commit will change the behaviour of unalias
when a not existing aliasname is given. The command
will newly return CMD_FAILURE and print a failure message.
Existings aliases will be deleted.

When updating an existing alias with the command alias,
then the old data structure will only be freed when
str_dup did not fail. This will prevent the data
of being corrupted in case of a memory failure.